extract theme updates to HOC
