+++ src/components/Auth/AuthServiceWrapper.tsx
@@ -1,4 +1,3 @@
-import { createTheme, ThemeOptions, ThemeProvider, useTheme } from '@mui/material'
 import { ErrorDialog } from '@xylabs/sdk-react'
 
 import { AuthActionTypes, AuthServiceId, useAuthState } from '../../contexts'
@@ -13,40 +12,6 @@ const backPayload = {
 const AuthServiceWrapper: React.FC = () => {
   const { state: authState, dispatch: authDispatch } = useAuthState()
 
-  const baseTheme = useTheme()
-  const authThemeOptions: ThemeOptions = {
-    components: {
-      MuiButton: {
-        styleOverrides: {
-          root: {
-            paddingBottom: baseTheme.spacing(2),
-            paddingLeft: baseTheme.spacing(1),
-            paddingRight: baseTheme.spacing(1),
-            paddingTop: baseTheme.spacing(2),
-            width: '100%',
-          },
-        },
-      },
-      MuiTextField: {
-        styleOverrides: {
-          root: {
-            marginBottom: baseTheme.spacing(2),
-          },
-        },
-      },
-      MuiTypography: {
-        styleOverrides: {
-          h3: {
-            marginBottom: baseTheme.spacing(4),
-            marginTop: baseTheme.spacing(4),
-          },
-        },
-      },
-    },
-  }
-  const authTheme = createTheme({ ...baseTheme, ...authThemeOptions })
-  console.log(authTheme)
-
   const handleLogout = () => {
     authDispatch({ payload: {}, type: AuthActionTypes.Logout })
     authDispatch(backPayload)
@@ -55,7 +20,7 @@ const AuthServiceWrapper: React.FC = () => {
   const handleBack = () => authDispatch(backPayload)
 
   return (
-    <ThemeProvider theme={authTheme}>
+    <>
       <MapActiveAuthService authState={authState} handleBack={handleBack} isLoading={authState.isLoading} />
       <AuthFooter handleLogout={handleLogout} isLoggedIn={authState.isLoggedIn} />
       {authState.authError && (
@@ -66,7 +31,7 @@ const AuthServiceWrapper: React.FC = () => {
           onAction={() => authDispatch({ payload: { authError: undefined }, type: AuthActionTypes.UpdateAuthError })}
         />
       )}
-    </ThemeProvider>
+    </>
   )
 }
