+++ src/components/Auth/index.ts
@@ -1,1 +1,2 @@
 export * from './AuthServiceWrapper'
+export * from './AuthThemeExtender'
